\name{ecg}
\docType{data}
\alias{ecg}
\title{Electrocardiogram Data}
\description{
  Electrocardiogram measurements for an individual with arhythmia.
  Observations were made over an 11.37 second interval at a rate of 180
  samples per second. For more detailed information regarding the
  collection of this data, see Percival and Walden (2000), p. 125.
}
\usage{data(ecg)}
\format{A time series object containing 2048 observations.}
\source{http://www.ms.washington.edu/~s530/data.html}
\references{
  Percival, D. B. and A. T. Walden (2000) \emph{Wavelet Methods for Time
    Series Analysis}, Cambridge University Press, sec. 4.10.
}
\keyword{datasets}
