\name{nile}
\docType{data}
\alias{nile}
\title{Yearly Nile River Minima}
\description{
  Yearly minimum water level of the Nile River starting in 622 A.D. and
  ending in 1284 A.D. For more information see Percival and Walden
  (2000), p. 190.
}
\usage{data(nile)}
\format{A time series object containing 683 observations.}
\source{http://www.ms.washington.edu/~s530/data.html}
\references{
  Percival, D. B. and A. T. Walden (2000) \emph{Wavelet Methods for Time
    Series Analysis}, Cambridge University Press, sec. 5.9.
}
\keyword{datasets}
