\name{ocean}
\docType{data}
\alias{ocean}
\title{Vertical Ocean Sheer Measurements}
\description{
  Measurements of vertical ocean sheer, obtained by dropping an
  instrument veritcally into the ocean and recording observations at 0.1
  meter intervals.  Starting depth for the series is 350.0 meters and
  ending depth is 1037.4 meters. For more detailed information regarding
  the collection of this data, see Percival and Walden (2000), p. 193.
}
\usage{data(ocean)}
\format{A time series object containing 6875 observations.}
\source{http://www.ms.washington.edu/~s530/data.html}
\references{
  Percival, D. B. and A. T. Walden (2000) \emph{Wavelet Methods for Time
    Series Analysis}, Cambridge University Press, sec. 5.10.
}
\keyword{datasets}
