\name{subtidal}
\docType{data}
\alias{subtidal}
\title{Subtidal Sea Level Measurements}
\description{
  Wwater level measurements from a stilling well on the open coast in
  Crescent Bay, CA. Observations were recorded every six minutes,
  from the beginning of 1980 to the end of 1991. For more detailed
  information regarding the collection of this data, see Percival and
  Walden (2000), p. 185.
}
\usage{data(subtidal)}
\format{A time series object containing 8746 observations.}
\source{http://www.ms.washington.edu/~s530/data.html}
\references{
  Percival, D. B. and A. T. Walden (2000) \emph{Wavelet Methods for Time
    Series Analysis}, Cambridge University Press, sec. 5.8.
}
\keyword{datasets}
